+++ src/main/java/com/ripple/BE/news/persistence/news/NewsRepository.java
@@ -1,4 +1,4 @@
-package com.ripple.BE.news.repository.news;
+package com.ripple.BE.news.persistence.news;
 
 import com.ripple.BE.news.domain.News;
 import jakarta.persistence.LockModeType;
