+++ src/main/java/com/ripple/BE/news/persistence/newscrap/NewsScrapRepository.java
@@ -1,4 +1,4 @@
-package com.ripple.BE.news.repository.newscrap;
+package com.ripple.BE.news.persistence.newscrap;
 
 import com.ripple.BE.news.domain.NewsScrap;
 import java.util.Optional;
