+++ src/main/java/com/ripple/BE/news/service/NewsService.java
@@ -10,9 +10,9 @@ import com.ripple.BE.news.domain.type.NewsSort;
 import com.ripple.BE.news.dto.NewsDTO;
 import com.ripple.BE.news.dto.NewsListDTO;
 import com.ripple.BE.news.exception.NewsException;
-import com.ripple.BE.news.repository.news.NewsJdbcRepository;
-import com.ripple.BE.news.repository.news.NewsRepository;
-import com.ripple.BE.news.repository.newscrap.NewsScrapRepository;
+import com.ripple.BE.news.persistence.news.NewsJdbcRepository;
+import com.ripple.BE.news.persistence.news.NewsRepository;
+import com.ripple.BE.news.persistence.newscrap.NewsScrapRepository;
 import com.ripple.BE.user.domain.User;
 import com.ripple.BE.user.service.AttendanceService;
 import com.ripple.BE.user.service.UserService;
